backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.5.0
faiss-cpu>=1.9.0
numpy>=1.24.0
openai>=1.3.0
web3>=6.11.0
eth-account>=0.9.0
aiofiles>=23.2.0
python-multipart>=0.0.6
beautifulsoup4>=4.12.0
lxml>=4.9.3